- Updated to Stackless Python 2.5.2 from http://svn.python.org/projects/stackless/branches/release25-maint revision 61023.

+++ branches/pspstackless-25-maint/Stackless/core/cframeobject.c
@@ -170,7 +170,7 @@ cframe_setstate(PyObject *self, PyObject *args)
 	PyObject *exec_name = NULL;
 	PyFrame_ExecFunc *good_func, *bad_func;
 	PyObject *params;
-	int i, n;
+	long i, n;
 
         if (!PyArg_ParseTuple (args, cframetuplenewfmt,
 			       &valid,
